+++ apps/mobile/src/app/(tabs)/settings/playback.tsx
@@ -216,6 +216,33 @@ export default function PlaybackSettingsPage() {
 						/>
 						<Checkbox.Item
 							mode='ios'
+							label='酷我音乐'
+							status={lyricSource === 'kuwo' ? 'checked' : 'unchecked'}
+							onPress={() => {
+								setSettings({ lyricSource: 'kuwo' })
+								setLyricSourceMenuVisible(false)
+							}}
+						/>
+						<Checkbox.Item
+							mode='ios'
+							label='酷狗音乐'
+							status={lyricSource === 'kugou' ? 'checked' : 'unchecked'}
+							onPress={() => {
+								setSettings({ lyricSource: 'kugou' })
+								setLyricSourceMenuVisible(false)
+							}}
+						/>
+						<Checkbox.Item
+							mode='ios'
+							label='百度音乐'
+							status={lyricSource === 'baidu' ? 'checked' : 'unchecked'}
+							onPress={() => {
+								setSettings({ lyricSource: 'baidu' })
+								setLyricSourceMenuVisible(false)
+							}}
+						/>
+						<Checkbox.Item
+							mode='ios'
 							label='自动 (选择最先返回的数据源)'
 							status={lyricSource === 'auto' ? 'checked' : 'unchecked'}
 							onPress={() => {