+++ packages/external-playlist-sync/src/types/common.ts
@@ -1,0 +1,6 @@
+export interface GenericTrack {
+	title: string
+	artists: string[]
+	album: string
+	duration: number // milliseconds
+}